Allow linking titlebar colours for Win95-style titlebars
Add a checkbox that when unchecked, disables setting the titlebar
gradient colours independently. When one end changes, the other is set
to the same colour. Active and inactive colours can still be set
independently.

To accommodate the new checkbox, limit the block display of labels to
just the fieldsets.

Breaking change: WinClassicTheme.prototype.enableLinkedElements now
takes an object of type Record<'button' | 'titlebar', boolean> (as expressed in TypeScript) instead of just a boolean.

Closes #11